+++ Kokoro/docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Proxied hooks are now marked with a proxy interface, allowing better integration.
+
 ## 1.6.0
 Released 18 April 2024.